Use focusInline from getSelection(), not value, when blocking select
At the time `onSelect` is called on a plugin, `change.value.selection`
hasn't been updated with the new selection. Since we block onSelect
based on `change.value.selection.focusInline`, this means that you can
end up blocking selection incorrectly in the following case:

1. Select an inline that's 3 characters long
2. Select character 3 of any other node
3. Watch as change.value.selection isn't updated with the new cursor position

This change relies on `window.getSelection()` for every check, and
double-checks `focusInline.key`, ensuring that we'll update the
selection when clicking on a different node.